first commit to modularizing code
